Organized the code.
Rename some of the files as to make their purpose clear.
