+++ shared_objects/include/compartment_per_object.h
