+++ shared_objects/include/static_variable.h
